requirements.txt
torch
torchvision
numpy
pandas
Pillow
tqdm
matplotlib
